Make "Enter" keypress on modal simulate Add click
Previously, pressing enter immediately set the
modal's returnValue to "add" which triggered the
creation of the card upon closure of modal. But it
did not account for the validity of the input.

I set the "Enter" keypress to simulate an Add
button click for it to go through the same logic,
thereby enforcing input validation.